Expose fetchResults function
The UI code does an admirable job of trying to cater for major frameworks
and the way that they do AJAX, but there are some cases that it just cannot
handle. For these cases, we expose the fetchResults() function and allow
applications to get their data however they are getting it, then call the
UI code with the UUID to fetch.